+++ src/index.js
@@ -1,5 +1,10 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
+import React from 'react'
+import ReactDOM from 'react-dom'
+import App from './App'
 
-ReactDOM.render(<App />, document.getElementById("root"));
+ReactDOM.render(<App />, document.getElementById('root'))
+
+setTimeout(() => {
+  console.log('making ajax call')
+}, 2000)
+console.log('scheduled ajax call, delay 2 seconds')
