+++ src/utils.js
@@ -1,0 +1,8 @@
+export const toggleOneTodo = (todos, index) => {
+  const newTodos = [...todos]
+  const todoToFlip = newTodos[index]
+  if (todoToFlip) {
+    todoToFlip.isCompleted = !todoToFlip.isCompleted
+  }
+  return newTodos
+}
